FPTS-10: Modularized display_and_data_handler.py & data_utils.py
display_and_data_handler.py -> tap_event_handler.py
Added: dynamodb_controller.py

data_utils.py -> file_utils.py + dynamodb_utils.py
